requirements.txt
google-adk>=0.1.0
mcp>=1.0.0
psycopg2-binary>=2.9.9
pytest>=7.0.0